Fix array attribute bug (#14)
* Fix stub library chunking + update test

* Move array attribute calls outside of loop

* Fix atomic tests

* Move the attribute prefix outside of the outlined procedure

* Add enum test

* Prevent type 'Range attributes from being rewritten

I found that my previous change broke 'Range attribute references when they were a part of types inside packages. For example, My_Package.My_Type'Range would break it because it saw that the prefix was not a simple identifier. This commit fixes that by checking if the prefix is a type.